New update

+++ recipes/nodemanager_mode.rb
@@ -40,6 +40,9 @@ search(:node, "role:hadoop-resourcemanager AND project:#{node['project']}").each
         node.set['hadoop']['yarn_site']['yarn.resourcemanager.webapp.https.address'] = rm + ":8090"
     end
 end
+
+node.save
+
 if ( File.exists?("/run/hadoop-yarn/yarn-yarn-nodemanager.pid") or node['hadoop_services']['already_nodemanager'] ) then
 
     Chef::Log.warn ("!!! U already have NodeManager running !!!")
@@ -59,47 +62,54 @@ if ( File.exists?("/run/hadoop-yarn/yarn-yarn-nodemanager.pid") or node['hadoop_
 
 else 
 
-    include_recipe "java_wrapper"
+    search(:node, "project:#{node['project']} AND role:hadoop-namenode AND hadoop_services_is_active_nn:true") do |m|
+        if ( m['hadoop_services']['already_resourcemanager'] ) then
 
-    include_recipe "hadoop::hadoop_yarn_nodemanager"
-           
-    tmp_hosts = node['hadoop']['hdfs_site']['dfs.namenode.http-address.mycluster.nn1']
-    dns_name = tmp_hosts[0..-7]
-    id_addr = Resolv.getaddress(dns_name)
+            include_recipe "java_wrapper"
 
-    ruby_block "add_to_hosts" do
-        block do
-            File.open('/etc/hosts', 'a') { |f| f.write("#{id_addr} mycluster") }
+            include_recipe "hadoop::hadoop_yarn_nodemanager"
+                   
+            tmp_hosts = node['hadoop']['hdfs_site']['dfs.namenode.http-address.mycluster.nn1']
+            dns_name = tmp_hosts[0..-7]
+            id_addr = Resolv.getaddress(dns_name)
+
+            ruby_block "add_to_hosts" do
+                block do
+                    File.open('/etc/hosts', 'a') { |f| f.write("#{id_addr} mycluster") }
+                end
+                not_if { File.open('/etc/hosts').lines.any?{|line| line.include?("#{id_addr}")} }
+            end
+
+            execute "hdfs-chown-dirs" do
+            command <<-EOF 
+                chown -R hdfs:hdfs #{node['hadoop']['hadoop_env']['hadoop_prefix']}-* 
+                chown -R hdfs:hdfs #{node['hadoop']['conf_dir']} 
+                chown -R hdfs:hdfs /tmp/hadoop-*
+                chown -R hdfs:hdfs #{node['hadoop']['hdfs_site']['dfs.ha.fencing.ssh.dir']}
+                chown -R zookeeper:zookeeper #{node['zookeeper']['zoocfg']['dataLogDir']}
+                chown -R hdfs:hdfs #{node['hadoop']['core_site']['hadoop.tmp.dir']}
+                chown -R hdfs:hdfs #{node['hadoop']['hadoop_env']['hadoop_log_dir']}
+                chown -R hdfs:hdfs #{node['hadoop']['hadoop_env']['hadoop_mapred_home']}
+            EOF
+            action :run
+            group "root"
+            user "root"
+            end
+
+            rewind "service[hadoop-yarn-nodemanager]" do
+              supports :status => true, :start => true, :stop => true, :restart => true
+              action [:enable, :start]
+            end
+
+            ruby_block "report_nodemanager_status" do
+                block do
+                node.set['hadoop_services']['already_nodemanager'] = true
+                end
+                action :nothing
+            end 
+        else
+            Chef::Log.info("I'm not ready yet")
         end
-        not_if { File.open('/etc/hosts').lines.any?{|line| line.include?("#{id_addr}")} }
     end
-
-    execute "hdfs-chown-dirs" do
-    command <<-EOF 
-        chown -R hdfs:hdfs #{node['hadoop']['hadoop_env']['hadoop_prefix']}-* 
-        chown -R hdfs:hdfs #{node['hadoop']['conf_dir']} 
-        chown -R hdfs:hdfs /tmp/hadoop-*
-        chown -R hdfs:hdfs #{node['hadoop']['hdfs_site']['dfs.ha.fencing.ssh.dir']}
-        chown -R zookeeper:zookeeper #{node['zookeeper']['zoocfg']['dataLogDir']}
-        chown -R hdfs:hdfs #{node['hadoop']['core_site']['hadoop.tmp.dir']}
-        chown -R hdfs:hdfs #{node['hadoop']['hadoop_env']['hadoop_log_dir']}
-        chown -R hdfs:hdfs #{node['hadoop']['hadoop_env']['hadoop_mapred_home']}
-    EOF
-    action :run
-    group "root"
-    user "root"
-    end
-
-    rewind "service[hadoop-yarn-nodemanager]" do
-      supports :status => true, :start => true, :stop => true, :restart => true
-      action [:enable, :start]
-    end
-
-    ruby_block "report_nodemanager_status" do
-        block do
-        node.set['hadoop_services']['already_nodemanager'] = true
-        end
-        action :nothing
-    end    
 
 end
